More tweaks
I've verified that the timer service removes sufficiently old cache data and
that the data is unpacked again on request. The reclaim operation is audited.
I should probably audit the unpack a well, but haven't done that here. I'm
still hoping for a successful CI run to check cobertura coverage.