Convert TOC to use cache
PBENCH-1192

This has been on my wishlist for a while, but was blocked by not actually
having a usable cache. PR #3550 introduces a functioning (if minimal) cache
manager, and this PR layers on top of that.

Note that, despite #3550 introducing a live cache, this PR represents the
first actual use of the in-memory cache map, and some adjustments were
necessary to make it work outside of the unit test environment.