Make cache management more dynamic
PBENCH-1301

This could be refined, but I want to get it up for review. I'd love to get it
on the staging server for testing with more limited disk space, although I can
ask for 95% or 880Gib on my laptop and see it free cache. (I'm out all next
week, but I may spend a bit of time trying to clean this up: e.g., trying to
add some unit tests.)

First, this adds code to track the unpacked size of tarballs to help with
managing cache goals. This is based on a `du -s -B1` of the unpacked directory
tree, and stored as metadata.

When asked to unpack, it will check whether there's sufficient space on the
cache device, and if not it will reclaim cache with a goal sufficient to
accommodate the target dataset.

The `pbench-tree-manage` command can now target either % free or bytes free,
and the background timer job will attempt to free 20% of the drive every 4
hours instead of targeting "old" tarballs. (The `last_ref` timestamp is now
used only to sort the list of datasets with live cache on input to reclaim
so that we free the oldest first.)