Make cache management more dynamic (#3578)
* Make cache management more dynamic

PBENCH-1301

First, this adds code to track the unpacked size of tarballs to help with
managing cache goals. This is based on a `du -s -B1` of the unpacked directory
tree, and stored as metadata.

When asked to unpack, it will check whether there's sufficient space on the
cache device, and if not it will reclaim cache with a goal sufficient to
accommodate the target dataset.

The `pbench-tree-manage` command can now target either % free or bytes free,
and the background timer job will attempt to free 20% of the drive every 4
hours instead of targeting "old" tarballs. (The `last_ref` timestamp is now
used only to sort the list of datasets with live cache on input to reclaim
so that we free the oldest first.)