Handle missing Opensearch indices
PBENCH-1308

An Opensearch-based query like `/api/v1/datasets/{dataset}/values/iterations`
requires that the `{dataset}` have at least one `result-data-sample` indexed
document. However it doesn't actually check that this is true. This leads to
an Opensearch query with a bad URL fragment causing an Opensearch error we
report as an internal server error.

This change detects early that the dataset has no indices for the required
index root name, and fails with a meaningful message.