Clean up
This combines a few issues: first, I've wanted to filter based on the unpacked
tarball size, but some tarballs are beyond the range of the SQL `INTEGER` type
and cause SQL cast errors -- change the interpretation of the `int` filter and
sort type to `BigInteger`. Also cleans up the logging around retried Sync
transaction errors, only logging warnings when it can't determine that the
error is a PostgreSQL serialization error. (I hope: this is hard to provoke in
casual testing.) Finally, clean up the logging of cached unpacked size by
avoiding two separate logs (without dataset name) on unpack, and adding a log
of the final unpacked size when we compute it.