Unit testing & review comments
PBENCH-1315

The index-map fix was submitted without unit tests so we can get it deployed
to resolve the immediate PostgreSQL resource problems. This follow-on adds
unit testing (and some late review comments).

__NOTE__: this isn't complete, but it's far easier to get a coverage report
from the CI than locally, so I'm posting a draft, which I'll be updating as
I build test cases.