Just more hacking around ...
This'll give some coverage of all three cases, but all based on the theory of
"call Elasticsearch which returns nothing", which isn't realistic. I'm just
experimenting with what I can twist out of the `Common` baseclass here, which
isn't really suited for this API but has enough flexibility (I think) to "make
do".