Aiming for 100% coverage of `/datasets/{id}`
This proved an "interesting" journey. I found several returns that don't really
make sense: for example, the old delete/update logic reported a `Sync.update`
failure as `CONFLICT` rather than `INTERNAL_SERVER_ERROR`. The fixture we use
install the Elasticsearch mock response and call the API needs to know whether
to install the mock: it has to be right as either installing the mock and not
calling it or calling it without the mock will result in failure. The current
logic around installation of the mock proved difficult to reconcile with the
need to handle `INTERNAL_SERVER_ERROR` both *before* and *after* the call to
Elasticsearch. As a result, I ended up massively refactoring the `query_api`
fixture code to simplify the logic based on a three-state override (neutral,
force, or suppress) in the new `expect_call` parameter.