More cleanup & adjustment
I fixed up mocking to properly unit test the `GET`, and realized that in the
current simplistic form, the output wasn't very useful as it wouldn't capture
the Elasticsearch `_index` value. A more useful output for a general `GET
/datasets/<id>` would be a summary rather than returning full documents. So
I rewrote the code and adjusted the unit tests.

Incidentally, after the production server PostgreSQL recovery, I found that my
report generator was tripping over datasets which had no operational status: I
suspect this was due to some hole in intake (although I think that failure to
create the `UPLOAD` operational status should have failed the intake, I'm not
going to debug that today). I did, however, adjust the report generator to
detect and report this case gracefully instead of failing with an f-string
formatting error. I also added a summary line of rows, tables, and size for
the SQL report.