Unit testing & review comments (#3608)
* Unit testing & review comments

PBENCH-1315

The index-map fix was submitted without unit tests so we can get it deployed
to resolve the immediate PostgreSQL resource problems. This follow-on adds
unit testing (and some late review comments).

This proved an "interesting" journey. I found several returns that don't really
make sense: for example, the old delete/update logic reported a `Sync.update`
failure as `CONFLICT` rather than `INTERNAL_SERVER_ERROR`. The fixture we use
install the Elasticsearch mock response and call the API needs to know whether
to install the mock: it has to be right as either installing the mock and not
calling it or calling it without the mock will result in failure. The current
logic around installation of the mock proved difficult to reconcile with the
need to handle `INTERNAL_SERVER_ERROR` both *before* and *after* the call to
Elasticsearch. As a result, I ended up massively refactoring the `query_api`
fixture code to simplify the logic based on a three-state override (neutral,
force, or suppress) in the new `expect_call` parameter.

Incidentally, after the production server PostgreSQL recovery, I found that my
report generator was tripping over datasets which had no operational status: I
suspect this was due to some hole in intake (although I think that failure to
create the `UPLOAD` operational status should have failed the intake, I'm not
going to debug that today). I did, however, adjust the report generator to
detect and report this case gracefully instead of failing with an f-string
formatting error. I also added a summary line of rows, tables, and size for
the SQL report.