Track and report unpack performance
This is a thought experiment, based off our earlier discussion of tracking an
asynchronous unpacker. I thought it would be fun to see some information on
tarball compression ratios, as well as tracking the min/max unpack time in
accessible metadata rather than just in logs. I need to jog back to Horreum,
but before diving back into the pool of Java, I took a recreational break...

So I added a simple `server.unpack-perf` metadata, which is a JSON block like
`{"min": <seconds>, "max": <seconds>, "count": <unpack_count>}`, and then
played with the report generator to get some statistics. The sample below is
for a `runlocal`, with a few small-ish tarballs. The big catch in deploying
this would be that none of the existing datasets will have
`server.unpack-perf` until they're unpacked again, which definitely reduces
the usefulness of this thought experiment.

Nevertheless, I figured I might as well post it for consideration.

```
Cache report:
  5 datasets currently unpacked, consuming 51.7 MB
  8 datasets have been unpacked a total of 15 times
  The least recently used cache was referenced today, pbench-user-benchmark_example-vmstat_2018.10.24T14.38.18
  The most recently used cache was referenced today, uperf_rhel8.1_4.18.0-107.el8_snap4_25gb_virt_2019.06.21T01.28.57
  The smallest cache is 4.1 kB, nometadata
  The biggest cache is 19.6 MB, trafficgen_basic-forwarding-example_tg:trex-profile_pf:forwarding_test.json_ml:5_tt:bs__2019-08-27T14:58:38
  The worst compression ratio is 22.156%, uperf_rhel8.1_4.18.0-107.el8_snap4_25gb_virt_2019.06.21T01.28.57
  The best compression ratio is 96.834%, pbench-user-benchmark_example-vmstat_2018.10.24T14.38.18
  The fastest cache unpack is 0.013 seconds, nometadata
  The slowest cache unpack is 0.078 seconds, trafficgen_basic-forwarding-example_tg:trex-profile_pf:forwarding_test.json_ml:5_tt:bs__2019-08-27T14:58:38
  The fastest cache unpack streaming rate is 253.666 Mb/second, trafficgen_basic-forwarding-example_tg:trex-profile_pf:forwarding_test.json_ml:5_tt:bs__2019-08-27T14:58:38
  The slowest cache unpack streaming rate is 0.133 Mb/second, nometadata
```