Add a "reclaim" tweak
The `pbench-tree-manage` utility supports a deep `ARCHIVE` tree display and
also managed the periodic background cache reclamation. Curious after we saw
a failed reclaim, I wanted to play with it a bit and realized that it does a
full (`search`) discovery unconditionally. First, even with `--display` (which
now that we have a proper report generator is rarely necessary) we probably
can use the faster SQL discovery most of the time, although I added an option
to select the slower `--search` discovery. More importantly, though, the
cache reclaimer doesn't need a fully discovered cache manager since it takes
the short-cut of examining the `/srv/pbench/cache` tree directly: so we can
move the discovery into the `--display` path.