"Complete"
This adds repair for `server.deletion` (expiration timestamp), completing the
repair of the `server` namespace.

In copying the setup from `intake_base.py` I realized that intake was
technically incorrect (not that it really matters much as we don't, and likely
won't, implement dataset expiration) in that it always uses the static
lifetime setting from `pbench-config.cfg` rather than recognizing the dynamic
server settings value. So I fixed that and made a common implementation.

It's also been bothering me that, in the midst of our PostgreSQL problems, we
allowed upload of datasets without metadata. I'd initially deliberatedly
allowed this looking at the metadata as "extra" and figuring I didn't want
to fail an upload just because of that. However, with recent optimizations,
we really depend internally on `server.tarball-path` in particular: the new
optimized `CacheManager.find_dataset` won't work without it. So failure in
setting metadata on intake is now a fatal internal server error.