A "re-imagining" of the old `pbench-reindex`
This supports deleting a datasets's indexed data and triggering re-indexing.
It also provides a mechanism to globally enable or disable indexing.

In the short term, this provides a workaround for our sharding issues, which
are preventing us from generating useful indexing anyway, by turning it off
entirely for new datasets. We can also delete indexed data (and the SQL
index maps) for existing datasets and, on demand, trigger re-indexing for
selected datasets.