Be more surer
I realized that even though I'd shortened the traceback, it would still be
displayed by Python on termination with an unhandled exception, and that would
still be relayed through conmon one line at a time. To avoid that, I added an
`except Exception` clause to print a simple message and exit.

In the process, I realized that while I'd written a new Click-based
`pbench-reindex`, the old `/opt/pbench-server/bin/pbench-reindex` script was
still there, and referenced in several config files. I removed these.