Fix deadlock and DELETE/UPDATE unindexed datasets
PBENCH-1328

A bug revealed another bug:

1. `POST`/`DELETE` `/datasets/{id}` fails when the dataset isn't indexed; with
the ability to disable server indexing entirely and to remove dataset indexed
data, this is undesirable. As long as the dataset doesn't have a `WORKING` ops
status, we should be able to update or delete.
2. When the operation fails, the status is already set to `WORKING`, and this
was not corrected on exit, leaving the dataset locked perpetually.

Fixed in two phases: first to correct the handling of errors so that we'll
always unlock on failure; second to allow deletion of unindexed datasets. (For
PR review, these phases are available in separate commits.)